Rename rogue as gateway
